requirements.txt
gTTS
playsound==1.2.2
python-slugify==6.1.2